Use own logger in mixin instead of shadowing, fixing 22w03a
With 22w03a's switch to slf4j, the shadowed field with that signature no longer exists. This means that ARRP crashes with mixin failures on 22w03a. This changes the mixin to use its own logger instead of relying on the shadowed one, so that old compat can be kept whilst fixing 22w03a.